+++ R/RunModule.R
@@ -1,4 +1,4 @@
-# Copyright 2022 Observational Health Data Sciences and Informatics
+# Copyright 2023 Observational Health Data Sciences and Informatics
 #
 # This file is part of Strategus
 #
@@ -21,8 +21,8 @@
 # carefully consider serialization and deserialization to JSON, which currently
 # uses custom functionality in ParallelLogger to maintain object attributes.
 
-runModule <- function(analysisSpecifications, moduleIndex, executionSettings, ...) {
-
+runModule <- function(analysisSpecifications, keyringSettings, moduleIndex, executionSettings, ...) {
+  checkmate::assert_multi_class(x = executionSettings, classes = c("CdmExecutionSettings", "ResultsExecutionSettings"))
   moduleSpecification <- analysisSpecifications$moduleSpecifications[[moduleIndex]]
 
   module <- moduleSpecification$module
@@ -31,7 +31,7 @@ runModule <- function(analysisSpecifications, moduleIndex, executionSettings, ..
   remoteUsername <- moduleSpecification$remoteUsername
   moduleFolder <- ensureModuleInstantiated(module, version, remoteRepo, remoteUsername)
 
-   # Create job context
+  # Create job context
   moduleExecutionSettings <- executionSettings
   moduleExecutionSettings$workSubFolder <- file.path(executionSettings$workFolder, sprintf("%s_%d", module, moduleIndex))
   moduleExecutionSettings$resultsSubFolder <- file.path(executionSettings$resultsFolder, sprintf("%s_%d", module, moduleIndex))
@@ -42,43 +42,101 @@ runModule <- function(analysisSpecifications, moduleIndex, executionSettings, ..
   if (!dir.exists(moduleExecutionSettings$resultsSubFolder)) {
     dir.create(moduleExecutionSettings$resultsSubFolder, recursive = TRUE)
   }
-  jobContext <- list(sharedResources = analysisSpecifications$sharedResources,
-                     settings = moduleSpecification$settings,
-                     moduleExecutionSettings = moduleExecutionSettings)
-  jobContextFileName <- gsub("\\\\", "/", tempfile(fileext = ".rds"))
+  jobContext <- list(
+    sharedResources = analysisSpecifications$sharedResources,
+    settings = moduleSpecification$settings,
+    moduleExecutionSettings = moduleExecutionSettings,
+    keyringSettings = keyringSettings
+  )
+  jobContextFileName <- file.path(moduleExecutionSettings$workSubFolder, "jobContext.rds") # gsub("\\\\", "/", tempfile(fileext = ".rds"))
   saveRDS(jobContext, jobContextFileName)
-
 
   # Execute module using settings
   script <- "
     source('Main.R')
     jobContext <- readRDS(jobContextFileName)
-    connectionDetails <- keyring::key_get(jobContext$moduleExecutionSettings$connectionDetailsReference)
-    connectionDetails <- ParallelLogger::convertJsonToSettings(connectionDetails)
-    connectionDetails <- do.call(DatabaseConnector::createConnectionDetails, connectionDetails)
-    jobContext$moduleExecutionSettings$connectionDetails <- connectionDetails
+
+    unlockKeyring <- function(keyringName) {
+      # If the keyring is locked, unlock it, set the value and then re-lock it
+      keyringLocked <- keyring::keyring_is_locked(keyring = keyringName)
+      if (keyringLocked) {
+        keyring::keyring_unlock(keyring = keyringName, password = Sys.getenv('STRATEGUS_KEYRING_PASSWORD'))
+      }
+      return(keyringLocked)
+    }
+
+    # If the keyring is locked, unlock it, set the value and then re-lock it
+    keyringName <- jobContext$keyringSettings$keyringName
+    keyringLocked <- unlockKeyring(keyringName = keyringName)
+  "
+
+  # Set the connection information based on the type of execution being
+  # performed
+  if (is(executionSettings, "CdmExecutionSettings")) {
+    script <- paste0(
+      script,
+      "connectionDetails <- keyring::key_get(jobContext$moduleExecutionSettings$connectionDetailsReference, keyring = keyringName)
+       connectionDetails <- ParallelLogger::convertJsonToSettings(connectionDetails)
+       connectionDetails <- do.call(DatabaseConnector::createConnectionDetails, connectionDetails)
+       jobContext$moduleExecutionSettings$connectionDetails <- connectionDetails"
+    )
+  } else if (is(executionSettings, "ResultsExecutionSettings")) {
+    script <- paste0(
+      script,
+      "resultsConnectionDetails <- keyring::key_get(jobContext$moduleExecutionSettings$resultsConnectionDetailsReference, keyring = keyringName)
+       resultsConnectionDetails <- ParallelLogger::convertJsonToSettings(resultsConnectionDetails)
+       resultsConnectionDetails <- do.call(DatabaseConnector::createConnectionDetails, resultsConnectionDetails)
+       jobContext$moduleExecutionSettings$resultsConnectionDetails <- resultsConnectionDetails"
+    )
+  } else {
+    stop("Unhandled executionSettings class! Must be one of the following: CdmExecutionSettings, ResultsExecutionSettings")
+  }
+
+  script <- paste0(script, "
+    if (keyringLocked) {
+       keyring::keyring_lock(keyring = keyringName)
+    }
 
     ParallelLogger::addDefaultFileLogger(file.path(jobContext$moduleExecutionSettings$resultsSubFolder, 'log.txt'))
     ParallelLogger::addDefaultErrorReportLogger(file.path(jobContext$moduleExecutionSettings$resultsSubFolder, 'errorReport.R'))
 
     options(andromedaTempFolder = file.path(jobContext$moduleExecutionSettings$workFolder, 'andromedaTemp'))
 
-    renv::use(lockfile = 'renv.lock')
+    if (Sys.getenv('FORCE_RENV_USE', '') == 'TRUE') {
+      renv::use(lockfile = 'renv.lock')
+    }
     execute(jobContext)
 
     ParallelLogger::unregisterLogger('DEFAULT_FILE_LOGGER', silent = TRUE)
     ParallelLogger::unregisterLogger('DEFAULT_ERRORREPORT_LOGGER', silent = TRUE)
-  "
+    writeLines('done', file.path(jobContext$moduleExecutionSettings$resultsSubFolder, 'done'))
+    ")
+
   script <- gsub("jobContextFileName", sprintf("\"%s\"", jobContextFileName), script)
-  tempScriptFile <- tempfile(fileext = ".R")
-  fileConn<-file(tempScriptFile)
+  tempScriptFile <- jobContextFileName <- file.path(moduleExecutionSettings$workSubFolder, "StrategusScript.R") # gsub("\\\\", "/", tempfile(fileext = ".R"))
+  fileConn <- file(tempScriptFile)
   writeLines(script, fileConn)
   close(fileConn)
 
-  renv::run(script = tempScriptFile,
-            job = FALSE,
-            name = "Running module",
-            project = moduleFolder)
+  doneFile <- file.path(jobContext$moduleExecutionSettings$resultsSubFolder, "done")
+  if (file.exists(doneFile)) {
+    unlink(doneFile)
+  }
+  renv::run(
+    script = tempScriptFile,
+    job = FALSE,
+    name = "Running module",
+    project = moduleFolder
+  )
+  if (!file.exists(doneFile)) {
+    message <- paste(
+      "Module did not complete. To debug:",
+      sprintf("  rstudioapi::openProject('%s', newSession = TRUE)", moduleFolder),
+      sprintf("  file.edit('%s')", tempScriptFile),
+      sep = "\n"
+    )
+    stop(message)
+  }
 
   return(list(dummy = 123))
 }